fix: move dayjs and just-throttle to dependencies
