--- index.js
@@ -1,21 +1,0 @@
-import Datepicker from './src/lib/components/Datepicker.svelte';
-import InlineCalendar from './src/lib/components/InlineCalendar.svelte';
-import Popover from './src/lib/components/Popover.svelte';
-import Calendar from './src/lib/components/calendar/Calendar.svelte';
-import Crossfade from './src/lib/components/generic/crossfade/Crossfade.svelte';
-import CrossfadeProvider from './src/lib/components/generic/crossfade/CrossfadeProvider.svelte';
-import Swappable from './src/lib/components/generic/Swappable.svelte';
-import Scrollable from './src/lib/components/generic/Scrollable.svelte';
-import * as themes from './src/lib/config/theme';
-
-export {
-	Datepicker,
-	InlineCalendar,
-	Popover,
-	Calendar,
-	Crossfade,
-	CrossfadeProvider,
-	Scrollable,
-	Swappable,
-	themes
-};
