+++ package.json
@@ -1,11 +1,10 @@
 {
   "name": "svelte-calendar",
   "version": "3.0.17",
-  "main": "index.js",
+  "main": "src/lib/index.js",
   "scripts": {
     "dev": "svelte-kit dev",
     "package": "svelte-kit package",
-    "postpackage": "node scripts/postpackage",
     "prebuild": "rm -rf build",
     "build": "svelte-kit build",
     "postbuild": "touch docs/.nojekyll",
@@ -14,17 +13,20 @@
     "lint": "prettier --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
     "format": "prettier --write --plugin-search-dir=. ."
   },
+  "dependencies": {
+    "dayjs": "^1.10.6",
+    "just-throttle": "^2.3.1"
+  },
   "devDependencies": {
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.1",
     "@semantic-release/npm": "^7.1.3",
     "@sveltejs/adapter-static": "^1.0.0-next.19",
     "@sveltejs/kit": "^1.0.0-next.147",
-    "dayjs": "^1.10.6",
     "eslint": "^7.22.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-svelte3": "^3.2.0",
-    "just-throttle": "^2.3.1",
+    "p-limit": "^4.0.0",
     "prettier": "~2.2.1",
     "prettier-plugin-svelte": "^2.2.0",
     "prism-svelte": "^0.4.7",
