+++ src/lib/index.js
@@ -1,0 +1,21 @@
+import Datepicker from './components/Datepicker.svelte';
+import InlineCalendar from './components/InlineCalendar.svelte';
+import Popover from './components/Popover.svelte';
+import Calendar from './components/calendar/Calendar.svelte';
+import Crossfade from './components/generic/crossfade/Crossfade.svelte';
+import CrossfadeProvider from './components/generic/crossfade/CrossfadeProvider.svelte';
+import Swappable from './components/generic/Swappable.svelte';
+import Scrollable from './components/generic/Scrollable.svelte';
+import * as themes from './config/theme';
+
+export {
+	Datepicker,
+	InlineCalendar,
+	Popover,
+	Calendar,
+	Crossfade,
+	CrossfadeProvider,
+	Scrollable,
+	Swappable,
+	themes
+};
